Add ADFS SMLD Compact tool; remove NestedImageSource thread
The thread caused untold race condition problems, so I removed it
in favour of a less fervent method of performing writes on thread
at time of write or source release, with the former being dependent
on a write not having taken place in a while.

This does have the drawback that in an interruption, not all data
may have been written.

Just like a real hard drive I guess.