Refactored the translator loaders
This makes more sense, and allows the upcoming audio translator loader
to be added without being a pain in the backside.

Instead of 3 functions doing the same thing, there is now only one.

Also, internal translators (not in plugins) are now referenced in the
BuiltIns class, saving having to special case each non-plugin translator
in both the translator loader and the plugin loader.