Memory Leak investigation/fixing - Round 1
This fixes a number of minor memory leaks, as well as DataSource
releasing relating the DataSourceCollector.

The CRT Debug report still shows hundreds of leaked blocks, but I cannot
trace what these are from. A couple I recognise their nature, but I
can't find any instance of objects of their type being allocated in a
way that won't cause deallocation.

For the time being I will leave them as is.

Still to come in the resource fixing rounds:

* I suspect FileOps (particular the Install Op) creates FileSystem
  objects that it may not be releasing. Need to check that. The
  corresponding recursive-out step should fix that though.

* I think some of the translator windows (notably SCREEN content) are
  using the old "Viewer closed" messages to ask CFileViewer (or even
  worse, NUTS itself) to close them. These need replacing with
  "delete this" paradigms.